refactor editor
