+++ pkg/editor/editor.go
@@ -1,6 +1,7 @@
 package editor
 
 import (
+	"bufio"
 	"bytes"
 	"encoding/json"
 	"fmt"
@@ -8,6 +9,8 @@ import (
 	"github.com/sirupsen/logrus"
 	"io/ioutil"
 	"os"
+	"os/exec"
+	"strings"
 )
 
 const cancelMessage = "Edit cancelled, no changes made."
@@ -23,7 +26,7 @@ type OnSaveFunc func(modifiedContent string) ([]string, error)
 
 func Edit(content string, name string, isJson bool, onSave OnSaveFunc) error {
 
-	tempFilePath, err := CreateTempFile()
+	tempFilePath, err := createTempFile()
 	if err != nil {
 		return err
 	}
@@ -46,7 +49,7 @@ func Edit(content string, name string, isJson bool, onSave OnSaveFunc) error {
 			return err
 		}
 
-		err = OpenEditor(tempFilePath)
+		err = openEditor(tempFilePath)
 		if err != nil {
 			return err
 		}
@@ -62,7 +65,7 @@ func Edit(content string, name string, isJson bool, onSave OnSaveFunc) error {
 		}
 
 		if isJson {
-			originalHasChanges := HasContentChanged(originalContent, currentContent)
+			originalHasChanges := hasContentChanged(originalContent, currentContent)
 			if !originalHasChanges {
 				return errors.New(cancelMessage)
 			}
@@ -87,7 +90,61 @@ func Edit(content string, name string, isJson bool, onSave OnSaveFunc) error {
 	return nil
 }
 
-func HasContentChanged(original, edited string) bool {
+func openEditor(filename string) error {
+	const vi = "vim"
+	var editor = os.Getenv("EDITOR")
+	var editorParts []string
+	if editor == "" {
+		editor = vi
+	}
+	editorParts = strings.Split(editor, " ")
+	editorPath := editorParts[0]
+
+	path, err := exec.LookPath(editorPath)
+	if err != nil {
+		return errors.New("ERROR: Editor \"" + editorPath + "\" specified in environment variable $EDITOR is not a valid program")
+	}
+
+	editorParts[0] = path
+
+	var cmd *exec.Cmd
+	cmd = new(exec.Cmd)
+	cmd.Path = path
+	cmd.Args = append(editorParts, filename)
+
+	cmd.Stdin = os.Stdin
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	err = cmd.Start()
+	if err != nil {
+		return err
+	}
+	err = cmd.Wait()
+
+	return err
+}
+
+func createTempFile() (string, error) {
+	const tmpFilePrefix = ".ao_edit_file_"
+	var tmpDir = os.TempDir()
+	tmpFile, err := ioutil.TempFile(tmpDir, tmpFilePrefix)
+	if err != nil {
+		return "", errors.New("Unable to create temporary file: " + err.Error())
+	}
+	return tmpFile.Name(), nil
+}
+
+func prettyPrintJson(jsonString string) string {
+	var out bytes.Buffer
+	err := json.Indent(&out, []byte(jsonString), "", "  ")
+	if err != nil {
+		return jsonString
+	}
+	return out.String()
+}
+
+func hasContentChanged(original, edited string) bool {
 
 	orgBuffer := new(bytes.Buffer)
 	err := json.Compact(orgBuffer, []byte(original))
@@ -102,4 +159,31 @@ func HasContentChanged(original, edited string) bool {
 	}
 
 	return orgBuffer.String() != editBuffer.String()
+}
+
+func stripComments(content string) string {
+	scanner := bufio.NewScanner(strings.NewReader(content))
+	var newline = ""
+	var actualContent string
+	for scanner.Scan() {
+		line := scanner.Text()
+		trimmedLine := strings.TrimSpace(line)
+		if !strings.HasPrefix(trimmedLine, "#") {
+			actualContent += newline + line
+			newline = "\n"
+		}
+	}
+
+	return actualContent
+}
+
+func addErrorMessage(messages []string) string {
+	comments := "#\n# ERROR:\n"
+	for _, message := range messages {
+		for _, line := range strings.Split(message, "\n") {
+			comments += fmt.Sprintf("# %s\n", line)
+		}
+	}
+
+	return comments + "#\n"
 }
