+++ pkg/editor/editor_test.go
@@ -40,7 +40,7 @@ func TestHasContentChanged(t *testing.T) {
 	}
 
 	for _, test := range testCases {
-		assert.Equal(t, test.Expected, HasContentChanged(test.Original, test.Edited))
+		assert.Equal(t, test.Expected, hasContentChanged(test.Original, test.Edited))
 	}
 
 }
